requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
faiss-cpu
python-multipart
